Update bit sizes

+++ CSES/Introductory_Problems/TwoKnights.cc
@@ -1,14 +1,15 @@
+#include <cstdint>
 #include <iostream>
 
-using i64 = long long;
-using u64 = unsigned long long;
-using u32 = unsigned;
+using i64 = int64_t;
+using u64 = uint64_t;
+using u32 = uint32_t;
 using u128 = unsigned __int128;
 
 int main() {
   std::cin.tie(nullptr)->sync_with_stdio(false);
 
-  int n;
+  u64 n;
   std::cin >> n;
 
   for (u64 i = 1; i <= n; i++) {
